fix: collect all build info in one file
