+++ lib/rspack.js
@@ -1,6 +1,5 @@
-import { resolve, join } from "path";
+import { resolve } from "path";
 import { fileURLToPath } from "url";
-import { mkdir, writeFile } from "fs/promises";
 import { runCommand, dirExist } from "./utils.js";
 
 const rootDir = resolve(fileURLToPath(import.meta.url), "../..");
@@ -29,41 +28,6 @@ async function switchToBranch(remote, branch) {
 	}
 }
 
-async function getCommitSHA() {
-	let commitSHA;
-	await runCommand("git", ["rev-parse", "HEAD"], {
-		onData(stdout) {
-			commitSHA = stdout.toString().trim();
-		}
-	});
-	console.log("Current Commit SHA", commitSHA);
-	return commitSHA;
-}
-
-async function getShortCommitSHA() {
-	let shortCommitSHA;
-	await runCommand("git", ["rev-parse", "--short", "HEAD"], {
-		onData(stdout) {
-			shortCommitSHA = stdout.toString().trim();
-		}
-	});
-	console.log("Current Short Commit SHA:", shortCommitSHA);
-	return shortCommitSHA;
-}
-
-async function recordRspackBuildInfo() {
-	const [commitSHA, shortCommitSHA] = await Promise.all([getCommitSHA(), getShortCommitSHA()]);
-	const outputDir = resolve(rootDir, "output");
-	await mkdir(outputDir, { recursive: true });
-
-	const buildInfoFile = join(outputDir, "build-info.json");
-	const buildInfo = {
-		commitSHA,
-		shortCommitSHA,
-	};
-	await writeFile(buildInfoFile, JSON.stringify(buildInfo, null, 2));
-}
-
 export async function buildRspack(remote, branch) {
 	if (!(await dirExist(rspackDir))) {
 		await runCommand("git", ["clone", repoUrl, ".rspack"]);
@@ -80,7 +44,4 @@ export async function buildRspack(remote, branch) {
 
 	console.log("== build rspack ==");
 	await runCommand("npm", ["run", "build:cli:release"]);
-
-	console.log("== recording rspack build info ==");
-	await recordRspackBuildInfo();
 }
